+++ util/steam.js
@@ -1,6 +1,12 @@
 const superagent = require('superagent');
 const { STEAM_API_KEY } = require('../config');
+const { Game } = require('../models');
 const { TAGS } = require('./constants');
+const { CHECK_MARK, CROSS_MARK } = require('../data/emojis.json')
+const { GREEN } = require('../data/colors.json');
+const { MessageEmbed } = require('discord.js');
+const { delay, crtHour } = require('../util/constants')
+const moment = require("moment");
 
 module.exports = client => {
     // TODO revoir exports, un steam.getGamesByName sera mieux qu'un client.getGamesByName
@@ -36,6 +42,40 @@ module.exports = client => {
 
         return search.body.items;
     };
+
+    client.getAllApps = async () => {
+        // https://api.steampowered.com/ISteamApps/GetAppList/v2/?key=xxx
+
+        const search = await superagent.get('https://api.steampowered.com/ISteamApps/GetAppList/v2/')
+                                        .query({
+                                            key: STEAM_API_KEY
+                                        });
+        return search.body?.applist?.apps
+    };
+
+    client.getCommunityApp = async appid => {
+        // https://api.steampowered.com/ICommunityService/GetApps/v1/?key=xxx&appids[0]=xxx
+        
+        const search = await superagent.get('https://api.steampowered.com/ICommunityService/GetApps/v1/')
+                                        .query({
+                                            key: STEAM_API_KEY,
+                                            appids: {appid},
+                                            language: 'fr'
+                                        })
+                                        .query(`appids[0]=${appid}`);
+        return search.body?.response?.apps
+    }
+
+    client.getAchievements = async appid => {
+        // https://api.steampowered.com/ISteamUserStats/GetGlobalAchievementPercentagesForApp/v2/?key=xxx&gameid=xxx
+        // https://api.steampowered.com/ISteamUserStats/GetGlobalAchievementPercentagesForApp/v2/?key=FC01A70E34CC7AE7174C575FF8D8A07F&gameid=321040
+        const search = await superagent.get('https://api.steampowered.com/ISteamUserStats/GetGlobalAchievementPercentagesForApp/v2/')
+                                        .query({
+                                            key: STEAM_API_KEY,
+                                            gameid: appid
+                                        });
+        return search
+    }
 
    /**
     * Récupère 50K apps qui ne sont ni des DLC, ni des software, ni des vidéos et ni des hardware
@@ -77,11 +117,11 @@ module.exports = client => {
      */
     client.getAppDetails = async appid => {
         const response = await superagent.get('https://store.steampowered.com/api/appdetails/?')
-            .query({
-                key: STEAM_API_KEY,
-                appids: appid
-            });
-        
+        .query({
+            key: STEAM_API_KEY,
+            appids: appid
+        });
+
         return response;
     };
 
@@ -105,5 +145,151 @@ module.exports = client => {
                 steamids: userid
             });
         return reponse;
+    }
+
+    /**
+     * 
+     */
+    client.fetchGame = async (appId, tag, nameTmp) => {
+        const app = await client.getAppDetails(appId);
+        //console.log(app.body[appId]);
+
+        let gameName = '', type = '';
+        let isMulti = false, isCoop = false, hasAchievements = false, isRemoved = false;
+        let update = {};
+
+        if (!app?.body[appId]?.success) {
+            // - chercher autre part car peut etre jeu "removed"
+            const communitApps = await client.getCommunityApp(appId)
+            
+            if (communitApps[0]?.name) {
+                isRemoved = true;
+                gameName = communitApps[0]?.name;
+                try {
+                    await client.getAchievements(appId);
+                    
+                    type = 'game'
+                    hasAchievements = true;
+                } catch (err) {
+                    hasAchievements = false;
+                }
+            } else {
+                gameName = nameTmp;
+                type = 'unknown'
+                //throw 'Jeu introuvable !'
+            }
+        } else {
+            type = app.body[appId].data?.type
+            gameName = app.body[appId].data?.name
+            let tags = app.body[appId].data?.categories
+            // au cas où pas de tags ou undefined
+            tags = tags ? tags : [];
+            // on ne garde que les tags qui nous intéresse (MULTI, COOP et ACHIEVEMENTS)
+            // TODO voir pour faire autrement ? récupérer tous les tags peu importe et faire recherche sur les tags via Mongo ?
+            isMulti = tags.some(tag => tag.id === TAGS.MULTI.id);
+            isCoop = tags.some(tag => tag.id === TAGS.COOP.id);
+            hasAchievements = tags.some(tag => tag.id === TAGS.ACHIEVEMENTS.id);
+        }
+        
+        // TODO icon plutot que l'image ? -> recup via API..
+        const gameUrlHeader = `https://steamcdn-a.akamaihd.net/steam/apps/${appId}/header.jpg`;
+        
+        const query = { appid: appId };
+        update = {
+            name: gameName,
+            type: type,
+            isMulti: isMulti,
+            isCoop: isCoop,
+            hasAchievements: hasAchievements,
+            isRemoved: isRemoved
+        };
+        
+        // on update ou créé le jeu
+        await Game.findOneAndUpdate(query, update, { upsert: true });
+
+        const msgCustom = `'${type}' trouvé et mis à jour !`;
+
+        //createLogs(client, interaction.guildId, `${gameName}`, `${msgCustom}`, '', GREEN)
+
+        const embed = new MessageEmbed()
+            .setColor(GREEN)
+            .setTitle(gameName)
+            .setDescription(`${msgCustom}`)
+            .setThumbnail(gameUrlHeader)
+            .addFields(
+                { name: '🌐 Multi', value: isMulti ? CHECK_MARK : CROSS_MARK, inline: true },
+                { name: '🤝 Co-op', value: isCoop ? CHECK_MARK : CROSS_MARK, inline: true },
+                { name: '🏆 Succès', value: hasAchievements ? CHECK_MARK : CROSS_MARK, inline: true },
+                // TODO ajouter lien Steam, ASTATS, CME etc
+            )
+            .setFooter({ text: `par ${tag}`});
+        
+        if (isRemoved)
+            embed.addFields({ name: '🚫 Removed', value: CHECK_MARK })
+
+        return embed;
+    }
+
+    client.fetchAllApps = async () => {
+        let startTime = moment();
+        let crtIdx = 1, cptGame = 0;
+
+        let apps = await client.getAllApps();
+        console.log(`trouvé ${apps.length}`);
+        
+        // - recup tous les appids de la bdd
+        const appidsDB = await Game.distinct('appid')
+    
+        // - remove name empty 
+        apps = apps.filter(item => item.name !== '')
+        console.log(`aftr name empty ${apps.length}`);
+        
+        // - garde seulement les appid divisible par 10 (autres ne sont pas des jeux)
+        // TODO je crois ?
+        apps = apps.filter(item => item.appid % 10 === 0)
+        console.log(`aftr mod 10 ${apps.length}`);
+        
+        // - remove appids déjà dans la bdd
+        // TODO a décommenter
+        apps = apps.filter(item => !appidsDB.includes(item.appid))
+        console.log(`aftr distinct ${apps.length}`);
+    
+        // TODO remove encore d'autres ?
+    
+        for (let i = 0; i < apps.length; i++) {
+            if (crtIdx % 250 === 0) {
+                logger.info("[" + crtHour() + "] - " + (crtIdx/apps.length) + " ..");
+                //await msgProgress.edit(`[${crtIdx}/${apps.length}] - Traitement des jeux ${".".repeat(((crtIdx/250) % 3) + 1)}`);
+            }
+
+            const app = apps[i];
+            cptGame++;
+            
+            console.log(` * go ${app.appid} ${app.name}`);
+            try {
+                await client.fetchGame(app.appid, 'system', app.name);
+            } catch (err) {
+                console.log('nope ' + app.name);
+
+                if (err.status === 429) {
+                    logger.info("\x1b[34m[INFO]\x1b[0m ["+crtHour()+"] - "+err+", on attend 5 min ..");
+                    //await msgProgress.edit(`${crtIdx}/${apps.length} - Trop de requêtes vers l'API Steam ! On attends 5 min ⏳`);
+                    // att 5 min
+                    await delay(300000);
+
+                    // on re essaie 
+                    try {
+                        await client.fetchGame(app.appid, 'system', app.name);
+                    } catch (error) {
+                        logger.error(`Ca veut pas, on peut rien faire pour ${app.name}...`);
+                    }
+                }
+            }
+
+            crtIdx++;
+        }
+
+        logger.info(".. Fin refresh games, "+cptGame+" jeux ajoutés");
+        return `Import des jeux terminés, ${cptGame} jeux ajoutés 👏`
     }
 }